requirements.txt
pydantic-settings>=2.0.3
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
sqlalchemy
pymysql
python-dotenv
opencv-python-headless
pytesseract
numpy
python-jose[cryptography]
pydantic-settings